shift the get all branches of repo and get all commits of all branch functions from utils files to each services files
