+++ services/bitbucketServices.js
@@ -59,8 +59,104 @@ const getRepositoryFromBitbucket = async (project) => {
     return false;
   }
 };
+//function for get all commits of all branches
+const getCommitsByBranches = async (repo, repoUrlInfo, branches) => {
+  let commitsObj = {};
+  let data = await branches.map(async (branch) => {
+    //can we just hit this API for master ,staging and production branches
+    try {
+      let isIncompleteCommits = true;
+      let allCommits = [];
+      let url = `https://api.bitbucket.org/2.0/repositories/${repoUrlInfo.handle}/${repoUrlInfo.repositorieName}/commits/${branch.name}?access_token=${process.env.BITBUCKET_ACCESS_TOKEN}`;
+      try {
+        while (isIncompleteCommits) {
+          const commits = await request.get(url);
+          if (repo.reviewed_at && commits.body.values.length > 0) {
+            if (
+              new Date(commits.body.values[0].date) > new Date(repo.reviewed_at)
+            ) {
+              if (
+                new Date(
+                  commits.body.values[commits.body.values.length - 1].date
+                ) > new Date(repo.reviewed_at)
+              ) {
+                allCommits = allCommits.concat(commits.body.values);
+                url = commits.body.next;
+              } else {
+                commits.body.values.map((commit) => {
+                  if (new Date(commit.date) >= new Date(repo.reviewed_at)) {
+                    allCommits.push(commit);
+                  }
+                });
+                url = commits.body.next;
+              }
+              if (!url) {
+                isIncompleteCommits = false;
+              }
+            } else {
+              isIncompleteCommits = false;
+            }
+          } else {
+            isIncompleteCommits = false;
+          }
+        }
+      } catch (err) {
+        logger.error(err);
+        logger.info("=========================================");
+      }
+      commitsObj[branch.name] = allCommits;
+    } catch (err) {
+      Sentry.captureException(err);
+      logger.error(
+        "Error executing while get all branches of  bitbucket repository in get all commits of each branches of repository function"
+      );
+      logger.error(err);
+      logger.info("=========================================");
+      commitsObj[branch.name] = false;
+    }
+  });
+  await Promise.all(data);
+  return commitsObj;
+};
+
+//function for get all branches of single repository
+const getAllBranchesOfRepo = async (repoInfo) => {
+  try {
+    let isIncompleteCommits = true;
+    let count = 10;
+    let allBranches = [];
+    let url = `https://api.bitbucket.org/2.0/repositories/${repoInfo.handle}/${repoInfo.repositorieName}/refs/branches?access_token=${process.env.BITBUCKET_ACCESS_TOKEN}&pagelen=10`;
+    while (isIncompleteCommits) {
+      let ProjectBranches = await request.get(url);
+      if (ProjectBranches.body) {
+        allBranches = allBranches.concat(ProjectBranches.body.values);
+        if (count < ProjectBranches.body.size) {
+          url = ProjectBranches.body.next;
+          count = count + 10;
+        } else {
+          break;
+        }
+      }
+    }
+    if (allBranches.length > 0) {
+      return allBranches;
+    } else {
+      return false;
+    }
+  } catch (err) {
+    Sentry.captureException(err);
+    logger.error(
+      "Error executing while get all branches of  bitbucket repository in get all branches of repository function"
+    );
+    logger.error(err);
+    logger.info("=========================================");
+    return null;
+  }
+};
 
 module.exports = {
   getTags: getTags,
   getRepositoryFromBitbucket: getRepositoryFromBitbucket,
+  getCommitsByBranches: getCommitsByBranches,
+  getAllBranchesOfRepo: getAllBranchesOfRepo,
 };
